added mapping from data model to return object

+++ src/main/VideoDB.WebApi/Models/ViewModels/RatingViewModel.cs
@@ -3,6 +3,6 @@ namespace Evo.WebApi.Models.ViewModels
     public class RatingViewModel
     {
         public string Source { get; set; }
-        public double RatingValue { get; set; }
+        public decimal RatingValue { get; set; }
     }
 }
